vagon refactoring

+++ ALARm_Report/Forms/PU32/LongPathProfilAndPlan.cs
@@ -297,7 +297,7 @@ namespace ALARm_Report.Forms
                                                                  track_profile.Select(p => p.Deviation).ToList(),
 
                                                                  //repers
-                                                                 RefPoints.Select(p => p.Mark).ToList(),
+                                                                 RefPoints.Select(p => (double)p.Mark).ToList(),
                                                                  x_RefPoints,
 
                                                                  yr,
