+++ ALARm_Report/Forms/longitudinalProfile/LongitudinalProfileElements.cs
@@ -197,7 +197,7 @@ namespace ALARm_Report.Forms
                                                                      track_profile.Select(p => p.Deviation).ToList(),
 
                                                                      //repers
-                                                                     RefPoints.Select(p => p.Mark).ToList(),
+                                                                     RefPoints.Select(p => (double)p.Mark).ToList(),
                                                                      x_RefPoints,
 
                                                                      yr,
